% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_TEs_mapped.R
\name{check_TEs_mapped}
\alias{check_TEs_mapped}
\title{generate region metadata from regions of interest given provided annotation files}
\usage{
check_TEs_mapped(
  bam,
  DogOverview = DogOverview,
  chrs = chrs,
  te_oi_locations = te_oi_locations
)
}
\arguments{
\item{bam}{bam filename}

\item{DogOverview}{dog_id metaData}

\item{chrs}{chromosomes vector}

\item{path_to_bams}{directory path}

\item{locations_oi}{locations of interest}
}
\value{
check a bam file for reads overlapping ones of interest
}
\description{
generate region metadata from regions of interest given provided annotation files
}
