% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatin_state_region_metaData_generation.R
\name{chromatin_state_region_metaData_generation}
\alias{chromatin_state_region_metaData_generation}
\title{generate chromatin region metadata from regions of interest given provided annotation file}
\usage{
chromatin_state_region_metaData_generation(
  regions,
  genome_gene_annotation,
  cpgisland_annotation
)
}
\arguments{
\item{regions}{regions of interest}

\item{chromatin_states_annotation}{chromatin states annotation}

\item{n.cores}{number of cores}
}
\value{
Function returns region metadata of interest
}
\description{
generate chromatin region metadata from regions of interest given provided annotation file
}
