% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_cgmaptoolsasm.R
\name{compile_cgmaptoolsasm}
\alias{compile_cgmaptoolsasm}
\title{compile the cgmaptools asm files}
\usage{
compile_cgmaptoolsasm(
  chr,
  path = "/scratch/bmarine2/meQTL-241004/traditional_asm_processing/"
)
}
\arguments{
\item{chr}{chromosome oi (e.g., 'chr1')}

\item{path}{path to asm files}
}
\value{
returns the compiled asm files for the chromosome of interest
}
\description{
compile the cgmaptools asm files
}
