% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_region_and_coverage.R
\name{filtering_region_and_coverage}
\alias{filtering_region_and_coverage}
\title{generate region, methylation, and coverage data of a bsseq object. can be run in parallel with a list of bsseq obects (e.g. by chromosomes)}
\usage{
filtering_region_and_coverage(dap_chr_oi, lower_bound = 0.1, upper_bound = 0.9)
}
\arguments{
\item{dap_chr_oi}{the bsseq object to be passed into the function-- can be done in parallel as a list with mclapply}

\item{lower_bound}{omit regions with less than X mean percent methylation}

\item{upper_bound}{omit regions with greater than X mean percent methylation}
}
\value{
Function returns region, methylation, and coverage information for the regions sought by regionFinder3
}
\description{
generate region, methylation, and coverage data of a bsseq object. can be run in parallel with a list of bsseq obects (e.g. by chromosomes)
}
