% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_data_for_plotting.R
\name{GO_data_for_plotting}
\alias{GO_data_for_plotting}
\title{generate region metadata from regions of interest given provided annotation files}
\usage{
GO_data_for_plotting(
  candidate_list,
  bg_genes,
  db = useMart("ENSEMBL_MART_ENSEMBL", dataset = "clfamiliaris_gene_ensembl", host =
    "https://www.ensembl.org"),
  go_ids = getBM(attributes = c("go_id", "external_gene_name", "namespace_1003"), filters
    = "external_gene_name", values = bg_genes, mart = db)
)
}
\arguments{
\item{candidate_list}{genes_of_interest}

\item{bg_genes}{background genes}

\item{db}{database}

\item{go_ids}{gene ids to be used}
}
\value{
Function returns dataframe of enriched gene ontology categories with their p-values
}
\description{
generate region metadata from regions of interest given provided annotation files
}
