% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_conserved_region_metaData_generation.R
\name{human_conserved_region_metaData_generation}
\alias{human_conserved_region_metaData_generation}
\title{generate transposable element region metadata from regions of interest given provided annotation files}
\usage{
human_conserved_region_metaData_generation(regions, conserved_annotation)
}
\arguments{
\item{regions}{regions of interest}

\item{conserved_annotation}{annotation file}
}
\value{
Function returns region metadata of interest
}
\description{
generate transposable element region metadata from regions of interest given provided annotation files
}
