% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMAGE_fnct.R
\name{IMAGE_fnct}
\alias{IMAGE_fnct}
\title{running image}
\usage{
IMAGE_fnct(
  format_for_image_chr,
  Kin = Kin <-
    readRDS("../BASELINE-PRECISION-240820/metaData_samples/240821-baseline_GRM.rds")
)
}
\arguments{
\item{format_for_image_chr}{the image formatted data}

\item{Kin}{kinship matrix}
}
\value{
returns image results
}
\description{
running image
}
