% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_hyper_hypo_plot.R
\name{odds_ratio_hyper_hypo_plot}
\alias{odds_ratio_hyper_hypo_plot}
\title{plot the odds ratio scatter plot for enriched and depleted regions from the hyper and hypo methylated regions of interest}
\usage{
odds_ratio_hyper_hypo_plot(
  pqlseq_res,
  region_metaData = region_metaData,
  fdr_perc = 0.05,
  omit_class = c(),
  label_order = c("Enriched in hyper & depleted in hypo", "Enriched in hyper only",
    "Depleted in hypo only",
    "No significance (p<sub>adj</sub> > 0.05, Fisher Exact test)",
    "Depleted in hyper only", "Enriched in hypo only",
    "Enriched in hypo & depleted in hyper", "Enriched in hyper & hypo",
    "Depleted in hyper & hypo")
)
}
\arguments{
\item{pqlseq_res}{pqlseq_results}

\item{region_metaData}{region metadata}

\item{fdr_perc}{fdr percent cutoff}

\item{omit_class}{what classes to omit from the scatterplot. e.g. c("SINE","LINE")}

\item{label_order}{the labels to be assigned to the plots in ggplot}
}
\value{
Function returns dataframe of oddsratio calculations for hypo and hyper comparisons of interest
}
\description{
plot the odds ratio scatter plot for enriched and depleted regions from the hyper and hypo methylated regions of interest
}
