% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQLseq_MW.R
\name{pqlseq_MW}
\alias{pqlseq_MW}
\title{modified PQLseq function by Marina Watowich to output covariate results
#### Updated MW Feb 11, 2019}
\usage{
pqlseq_MW(
  RawCountDataSet,
  Phenotypes = NULL,
  Covariates = NULL,
  RelatednessMatrix = NULL,
  LibSize = NULL,
  fit.model = "PMM",
  fit.method = "AI.REML",
  fit.maxiter = 500,
  fit.tol = 1e-05,
  numCore = 1,
  filtering = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{RawCountDataSet}{methylation mat}

\item{Covariates}{covariates mat}

\item{RelatednessMatrix}{Relatedness Matrix}

\item{numCore}{numer of cores}
}
\value{
BMM mixed-model result
}
\description{
modified PQLseq function by Marina Watowich to output covariate results
#### Updated MW Feb 11, 2019
}
