% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_plink.R
\name{read_in_plink_path}
\alias{read_in_plink_path}
\title{generate plink_in file}
\usage{
read_in_plink_path(chr)
}
\arguments{
\item{chr}{chromosome}
}
\value{
Function reads in plink
}
\description{
generate plink_in file
}
