% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_metaData_generation.R
\name{region_metaData_generation}
\alias{region_metaData_generation}
\title{generate region metadata from regions of interest given provided annotation files}
\usage{
region_metaData_generation(
  regions,
  n.cores,
  gtf =
    rtracklayer::import("../../GENOME-ANNOTATION-FILE/UU_Cfam_GSD_1.0_ROSY.refSeq.ensformat.gtf"),
  gtf_cpgisl = gtf_cpgisl,
  chromatin_state_bed_file = bed_file
)
}
\arguments{
\item{regions}{regions of interest}

\item{n.cores}{numer of cores}

\item{gtf}{gene annotation file}

\item{gtf_cpgisl}{CpG island annotation file}

\item{chromatin_state_bed_file}{bed file of the chromatin state annotation}
}
\value{
Function returns region metadata of interest
}
\description{
generate region metadata from regions of interest given provided annotation files
}
