% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_metaData_median_percent_methylation.R
\name{region_metaData_median_percent_methylation}
\alias{region_metaData_median_percent_methylation}
\title{returns a list of plots}
\usage{
region_metaData_median_percent_methylation(
  coverage_all_chr,
  methylation_all_chr,
  region_metaData,
  col_oi = c("Promoter", "exon", "intron", "upstream_utr", "downstram_utr", "CpG_shore",
    "CpG_island", "LINE", "SINE", "ActiveTSS", "WeakTSS", "FlankingActiveTSS1",
    "FlankingActiveTSS2", "ActiveStrongEnhancer", "ActiveWeakEnhancer",
    "ActivePoisedEnhancer", "BivalentTssEnh", "RepressedPolycomb", "Repressed",
    "ZNFgenesRepeats", "QuiescenatLow", "Heterochromatin")
)
}
\arguments{
\item{coverage_all_chr}{region coverage}

\item{methylation_all_chr}{region methylation}

\item{region_metaData}{region metaData}

\item{col_oi}{columns of interest of the region metaData}
}
\value{
Function returns a list of plots
}
\description{
returns a list of plots
}
