% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaData_add_shinyinfo.R
\name{shiny_meta_add}
\alias{shiny_meta_add}
\title{build sample metadata}
\usage{
shiny_meta_add(metaData, shiny_metaData, cols_oi = c("rdid", "prep_date"))
}
\arguments{
\item{metaData}{alpha to use for glmnet}

\item{shiny_metaData}{alpha to use for glmnet}

\item{cols_oi}{what cols to cbind to the metaData}
}
\value{
Function returns the metadata with columns from shiny
}
\description{
build sample metadata
}
