% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_asm.R
\name{sites_with_one_instance_of_asm}
\alias{sites_with_one_instance_of_asm}
\title{formatting for image}
\usage{
sites_with_one_instance_of_asm(asm_res_oi)
}
\arguments{
\item{asm_res_oi}{the .asm file that you'd like to filter}
}
\value{
returns the snp cpg combinations with allele-specific methylation
}
\description{
formatting for image
}
