% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_for_image_chr.R
\name{VectorIntersect}
\alias{VectorIntersect}
\title{formatting for image by chr}
\usage{
VectorIntersect(v, z)
}
\arguments{
\item{asm_chr}{the .asm file that you'd like to filter}
}
\value{
returns the image formated requirements
}
\description{
formatting for image by chr
}
