% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_clock.R
\name{build_clock}
\alias{build_clock}
\title{generate region metadata from regions of interest given provided annotation files}
\usage{
build_clock(
  alph,
  metaData,
  region_metaData,
  perc_meth,
  test_samples = colnames(perc_meth) \%in\% metaData$lid_pid[metaData$Cohort !=
    "precision_1"],
  seed = 100
)
}
\arguments{
\item{alph}{alpha to use for glmnet}

\item{metaData}{sample metaData}

\item{region_metaData}{region metaData}

\item{perc_meth}{imputed percent methylation matrix}

\item{test_samples}{test samples for elastic net}
}
\value{
Function returns a list of predicted age as a column to the input metaData, region weights as a column to the input region_metaData, and clock results as a list
}
\description{
generate region metadata from regions of interest given provided annotation files
}
