% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_TEs_mapped.R
\name{check_TEs_mapped}
\alias{check_TEs_mapped}
\title{check the reads that map to TEs in chromsomes of interest, and return the age and some metadata information for the dog}
\usage{
check_TEs_mapped(
  bam,
  DogOverview = DogOverview,
  chrs = chrs,
  te_oi_locations = te_oi_locations,
  path_to_bams = "/scratch/vsohrab/dap_map_canfam4/nvidia_fq2bam_mapping/bams/"
)
}
\arguments{
\item{bam}{bam filename}

\item{DogOverview}{dog_id metaData}

\item{chrs}{chromosomes vector}

\item{te_oi_locations}{locations of interest}

\item{path_to_bams}{directory path}
}
\value{
check a bam file for reads overlapping ones of interest
}
\description{
check the reads that map to TEs in chromsomes of interest, and return the age and some metadata information for the dog
}
