% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_clock.R
\name{clock_simulate}
\alias{clock_simulate}
\title{clock simulations}
\usage{
clock_simulate(
  alph,
  metaData,
  region_metaData,
  perc_meth,
  region_metaData_te,
  perc_meth_te,
  region_metaData_prmtr,
  perc_meth_prmtr,
  region_metaData_cpgisl,
  perc_meth_cpgisl
)
}
\arguments{
\item{metaData}{sample metaData}

\item{region_metaData}{region metaData}

\item{perc_meth}{imputed percent methylation matrix}

\item{region_metaData_te}{TE region metaData}

\item{perc_meth_te}{imputed TE percent methylation matrix}

\item{region_metaData_prmtr}{promoter region metaData}

\item{perc_meth_prmtr}{imputed promoter percent methylation matrix}

\item{region_metaData_cpgisl}{promoter region metaData}

\item{perc_meth_cpgisl}{imputed promoter percent methylation matrix}

\item{alphs_oi}{alphas to use for glmnet}
}
\value{
Function returns a list of the simulated clock results
}
\description{
clock simulations
}
