% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_hyper_hypo_plot.R
\name{odds_ratio_hyper_hypo_plot}
\alias{odds_ratio_hyper_hypo_plot}
\title{plot the odds ratio scatter plot for enriched and depleted regions from the hyper and hypo methylated regions of interest}
\usage{
odds_ratio_hyper_hypo_plot(
  pqlseq_res,
  region_metaData = region_metaData,
  fdr_perc = 0.05
)
}
\arguments{
\item{pqlseq_res}{pqlseq_results}

\item{region_metaData}{region metadata}

\item{fdr_perc}{fdr percent cutoff}
}
\value{
Function returns plot of annotations from the region_metaData that are enriched/depleted in hypomethyated and hypermethylated regions
}
\description{
plot the odds ratio scatter plot for enriched and depleted regions from the hyper and hypo methylated regions of interest
}
