% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_hyper_hypo_plot.R
\name{odds_ratio_hyper_hypo_plot}
\alias{odds_ratio_hyper_hypo_plot}
\title{plot the odds ratio scatter plot for enriched and depleted regions from the hyper and hypo methylated regions of interest}
\usage{
odds_ratio_hyper_hypo_plot(
  pqlseq_res,
  region_metaData = region_metaData,
  fdr_perc = 0.05,
  omit_class = c(),
  color_order = c("purple", "blue", "darkgreen", "black", "darkorange", "maroon", "red")
)
}
\arguments{
\item{pqlseq_res}{pqlseq_results}

\item{region_metaData}{region metadata}

\item{fdr_perc}{fdr percent cutoff}

\item{omit_class}{what classes to omit from the scatterplot. e.g. c("SINE","LINE")}

\item{color_order}{the colors to be assigned to the plots in ggplot}
}
\value{
Function returns plot of annotations from the region_metaData that are enriched/depleted in hypomethyated and hypermethylated regions
}
\description{
plot the odds ratio scatter plot for enriched and depleted regions from the hyper and hypo methylated regions of interest
}
