% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_metaData_generation.R
\name{region_metaData_generation}
\alias{region_metaData_generation}
\title{generate region metadata from regions of interest given provided annotation files}
\usage{
region_metaData_generation(
  regions,
  genome_gene_annotation,
  cpgisland_annotation,
  chromatin_states_annotation
)
}
\arguments{
\item{regions}{regions of interest}

\item{genome_gene_annotation}{gene annotation file}

\item{cpgisland_annotation}{CpG island annotation file}

\item{chromatin_states_annotation}{chromatin state annotation (not currently used)}

\item{n.cores}{numer of cores}
}
\value{
Function returns region metadata of interest
}
\description{
generate region metadata from regions of interest given provided annotation files
}
