% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_metaData_median_percent_methylation.R
\name{region_metaData_median_percent_methylation}
\alias{region_metaData_median_percent_methylation}
\title{returns a list of plots}
\usage{
region_metaData_median_percent_methylation(
  coverage_all_chr,
  methylation_all_chr,
  region_metaData,
  include_y_axis_labels = FALSE,
  col_oi = c("Promoter", "exon", "intron", "upstream_utr", "downstram_utr", "CpG_shore",
    "CpG_island", "LINE", "SINE", "ActiveTSS", "WeakTSS", "FlankingActiveTSS1",
    "FlankingActiveTSS2", "ActiveStrongEnhancer", "ActiveWeakEnhancer",
    "ActivePoisedEnhancer", "BivalentTssEnh", "RepressedPolycomb", "Repressed",
    "ZNFgenesRepeats", "QuiescenatLow", "Heterochromatin"),
  theme_blaise = theme(axis.text.x = element_text(angle = 0), plot.title =
    element_text(family = "sans", size = 24, hjust = 0.5, color = "black", face =
    "bold"), plot.subtitle = element_text(family = "sans", size = 11, color = "black"),
    axis.text = element_text(family = "sans", size = 18, color = "black"), axis.title =
    element_text(family = "sans", size = 20, color = "black"), panel.border =
    element_blank(), axis.line = element_line(colour = "black", linewidth = 1),
    axis.ticks = element_line(colour = "black", linewidth = 1), 
     legend.key.size =
    unit(1.5, "cm"), legend.key = element_rect(fill = NA), legend.text =
    element_text(family = "sans", size = 20), legend.title = element_blank(),
    legend.background = element_blank(), legend.box.background = element_blank(),
    legend.text.align = 0, panel.background = element_blank(), panel.grid.major =
    element_line(colour = "black"), panel.grid.minor = element_blank()) + removeGrid()
)
}
\arguments{
\item{coverage_all_chr}{region coverage}

\item{methylation_all_chr}{region methylation}

\item{region_metaData}{region metaData}

\item{include_y_axis_labels}{if TRUE will remove the y axis title}

\item{col_oi}{columns of interest of the region metaData}

\item{theme_blaise}{ggplot theme for plotting}
}
\value{
Function returns a list of plots
}
\description{
returns a list of plots
}
