% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_region_metaData_generation.R
\name{TE_region_metaData_generation}
\alias{TE_region_metaData_generation}
\title{generate region metadata from regions of interest given provided annotation files}
\usage{
TE_region_metaData_generation(
  regions,
  transposons_annotation,
  sw_score_cutoff = 0
)
}
\arguments{
\item{regions}{regions of interest}

\item{transposons_annotation}{TE annotation file}

\item{sw_score_cutoff}{the minimum sw score that will be queried here}
}
\value{
Function returns region metadata of interest
}
\description{
generate region metadata from regions of interest given provided annotation files
}
